optmethods: move testing code to test_opt_original.py
The optimisation code had "tests" which could be run by evaluating
the module directly (run via a "if __name__ == '__main__':" block).

These tests are not pytest-like tests, and instead run a number
of optimisation tests with a variety of methods, displaying the
results. These were presumably used to validate and verify changes
made to the code (as there is no way to impove the results for all
cases).

One problem with this arrangement is that there was a lot of code
in the modules that were nothing to do with the optimiser, and it
makes it hard to see what is tested and what is relevant. So, this
commit moves all the "testing" code into test_opt_original.py
and makes three of the cases actual tests (those from opt.py).
The code from ncoresde.py and ncoresnm.py are left "as is", and
can only be run by executing the file directly. They do not
actually test anything, just output the results. For example:

% python sherpa/optmethods/tests/test_opt_original.py --no_de

 rosenbrock  fmin = 0.0
ncoresNelderMead [0.99999986 0.99999963 1.00000032 1.00000065] = 9.690772711926466e-13 in 2419 nfevs

 freudenstein_roth  fmin = 0.0
ncoresNelderMead [11.41277901 -0.89680525 11.41277899 -0.89680526] = 97.96850735848004 in 1744 nfevs

 powell_badly_scaled  fmin = 0.0
ncoresNelderMead [1.09822720e-05 9.10558361e+00 1.09822720e-05 9.10558361e+00] = 1.0696645337509389e-14 in 7790 nfevs

...

 SumSquares(0, 0, ..., 0) = 0
ncoresNelderMead [0. 0. 0. 0.]  =  0.0 in 1466 nfevs

 Zakharov(0, 0, ..., 0) = 0
ncoresNelderMead [0. 0. 0. 0. 0. 0. 0. 0.]  =  0.0 in 4701 nfevs

And here's the full list of options (cobbled together from the
existing ncoresnm/de code; I have not changed things so I think
the -s flag is actually the inverse of the suggested text):

% python sherpa/optmethods/tests/test_opt_original.py --help
usage: test_opt_original.py [-h] [-N NUM] [-u] [-o] [-d] [-c] [-s] [--no_de] [--no_nm]

options:
  -h, --help     show this help message and exit
  -N NUM
  -u, --unc_opt  do not run tst_unc_opt
  -o, --opt      do not run tst_opt
  -d, --difevo   run simple difevo (ncoresde)
  -c, --combine  run nm & difevo (ncoresde)
  -s, --single   run test using 1 core (ncoresnm)
  --no_de        skip ncoresde tests
  --no_nm        skip ncoresnm tests

It would be nice to run this code as an actual test, since we can
at least check that things havent been changed **but**

- this is a lot of work
  - need to split up each test into an individual test rather than
    the current "run all"
  - need to identify what we can test
- it's not clear how much of the code that is run in these "tests"
  is actually used by Sherpa, and how much is experimental code
