+++ sherpa/optmethods/ncoresde.py
@@ -444,44 +444,3 @@ class ncoresDifEvoNelderMead:
             my_par = nm_par
 
         return nfev, my_fmin, my_par
-
-
-if '__main__' == __name__:
-
-    # from sherpa.optmethods.opt import tst_opt, tst_unc_opt
-    from opt import tst_opt, tst_unc_opt
-
-    import argparse
-    parser = argparse.ArgumentParser()
-    parser.add_argument('-d', '--difevo', action="store_true",
-                        default=False, help='run simple difevo', dest="difevo")
-    parser.add_argument('-c', '--combine', action="store_true",
-                        default=False, help='run nm & difevo', dest="combine")
-    parser.add_argument("-u", "--unc_opt", dest="unc_opt", default=True,
-                        action="store_false", help="do not run tst_unc_opt")
-    parser.add_argument("-o", "--opt", dest="global_func", default=True,
-                        action="store_false", help="do not run tst_opt")
-    parser.add_argument('-N', action="store", dest="num", default=4, type=int)
-
-    options = parser.parse_args()
-    if options.num % 2 != 0:
-        raise ValueError("-N option must be an even number")
-
-    if options.difevo:
-        algo = [DifEvo()]
-        if options.unc_opt:
-            tst_unc_opt(algo, options.num)
-        if options.global_func:
-            tst_opt(algo, options.num)
-    elif options.combine:
-        algo = [ncoresDifEvoNelderMead()]
-        if options.unc_opt:
-            tst_unc_opt(algo, options.num)
-        if options.global_func:
-            tst_opt(algo, options.num)
-    else:
-        algo = [ncoresDifEvo()]
-        if options.unc_opt:
-            tst_unc_opt(algo, options.num)
-        if options.global_func:
-            tst_opt(algo, options.num)
