+++ sherpa/optmethods/ncoresnm.py
@@ -403,40 +403,3 @@ class ncoresNelderMeadRecursive(ncoresNelderMead):
             return self.calc(fcn, par, xmin, xmax, tol, maxnfev, numcores, fmin, nfev)
 
         return nfev, fval, par
-
-
-if '__main__' == __name__:
-
-    from opt import tst_opt, tst_unc_opt
-
-    from argparse import ArgumentParser
-    parser = ArgumentParser()
-    parser.add_argument("-N", "--num", dest="num", default=10,
-                        type=int, help="set num")
-    parser.add_argument("-s", "--single", dest="single", default=True,
-                        action="store_false", help="run test using 1 core")
-    parser.add_argument("-u", "--unc_opt", dest="unc_opt", default=True,
-                        action="store_false", help="do not run tst_unc_opt")
-    parser.add_argument("-o", "--opt", dest="global_func", default=True,
-                        action="store_false", help="do not run tst_opt")
-    args = parser.parse_args()
-
-    if args.num % 2 != 0:
-        raise ValueError("-N option must be an even number")
-
-    if args.single:
-        nmcores = ncoresNelderMead()
-        # midnight = Midnight()
-        algorithms = [nmcores]
-        if args.unc_opt:
-            tst_unc_opt(algorithms, args.num)
-        if args.global_func:
-            tst_opt(algorithms, args.num)
-    else:
-        algorithms = [NelderMead0(), NelderMead1(), NelderMead2(),
-                      NelderMead3(), NelderMead4(), NelderMead5(),
-                      NelderMead6(), NelderMead7()]
-        if args.unc_opt:
-            tst_unc_opt(algorithms, args.num)
-        if args.global_func:
-            tst_opt(algorithms, args.num)
